+++ Studenda.Core/Data/DataContext.cs
@@ -4,8 +4,10 @@ using Studenda.Core.Data.Configuration;
 using Studenda.Core.Model;
 using Studenda.Core.Model.Common;
 using Studenda.Core.Model.Schedule;
+using Studenda.Core.Model.Schedule.Link;
 using Studenda.Core.Model.Schedule.Management;
 using Studenda.Core.Model.Security;
+using Studenda.Core.Model.Security.Link;
 using Studenda.Core.Model.Security.Management;
 
 namespace Studenda.Core.Data;
