+++ Studenda.Core/Studenda.Core.csproj
@@ -10,4 +10,6 @@
         <PackageReference Include="Microsoft.EntityFrameworkCore.Sqlite" Version="7.0.9"/>
     </ItemGroup>
 
+
+
 </Project>
